+++ tests/test_query_hazard_agg_vs30_fix.py
@@ -8,6 +8,8 @@ from nzshm_common.location.location import LOCATIONS_BY_ID
 
 from toshi_hazard_store import model, query_v3
 
+# import toshi_hazard_store.model.caching.cache_store
+
 HAZARD_MODEL_ID = 'MODEL_THE_FIRST'
 vs30s = [250, 500, 1000, 1500]
 imts = ['PGA']
@@ -30,7 +32,7 @@ def build_hazard_aggregation_models():
             ).set_location(loc)
 
 
-@patch("toshi_hazard_store.v2.db_adapter.sqlite.sqlite_store.LOCAL_CACHE_FOLDER", None)
+@patch("toshi_hazard_store.model.caching.cache_store.LOCAL_CACHE_FOLDER", None)
 @mock_dynamodb
 class QueryHazardAggregationV3TestVS30(unittest.TestCase):
     def setUp(self):